# API.yaml
openapi: 3.0.3
info:
  title: Printer Queue Generation API
  description: |
    API para geração de filas de impressão com empacotamento automático de imagens.
    
    ## Funcionalidades Principais
    
    - **Gestão de Tenants**: Multi-tenancy com isolamento de dados
    - **Gestão de Máquinas**: Configuração de máquinas de impressão por tenant
    - **Gestão de Assets**: Catálogo de imagens indexadas por SKU
    - **Gestão de Perfis de Dimensionamento**: Perfis para dimensionamento automático de imagens
    - **Gestão de Storage**: Configuração de storage (S3, Local, Dropbox)
    - **Processamento de Jobs**: Upload de picklists PDF, resolução de SKUs, empacotamento e geração de PDFs
    - **Image Packing**: Upload direto de imagens para empacotamento sem validação de SKU
    
    ## Autenticação
    
    A API usa multi-tenancy baseado em header `X-Tenant-ID`. Todos os endpoints (exceto `/healthz`) requerem este header.
    
    ## Status de Jobs
    
    - `queued`: Job criado, aguardando processamento
    - `processing`: Job sendo processado pelo worker
    - `needs_input`: Job precisa de resolução manual de SKUs
    - `completed`: Job processado com sucesso
    - `failed`: Job falhou durante processamento
    
    ## Modos de Processamento
    
    - `sequence`: Processa itens na ordem exata do picklist PDF
    - `optimize`: Otimiza layout para minimizar número de bases e desperdício
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
...
paths:
...
  /v1/jobs:
    post:
      tags:
        - jobs
      summary: Upload picklist PDF and create job
      description: |
        Upload a picklist PDF file and create a new job for processing.
        
        **Business Rules:**
        - File must be PDF format
        - Maximum file size: 10MB
        - Job is created with status 'queued'
        - Worker task is automatically enqueued
        - Mode 'sequence' processes items in PDF order
        - Mode 'optimize' optimizes layout to minimize bases and waste
        
        **Processing Pipeline:**
        1. Parse PDF to extract SKUs and quantities
        2. Resolve SKUs against asset catalog
        3. If unresolved SKUs exist, job status becomes 'needs_input'
        4. Apply sizing profiles to items
        5. Pack items into bases using selected mode
        6. Generate print-ready PDFs for each base
      operationId: uploadPicklist
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Picklist PDF file (max 10MB)
                mode:
                  type: string
                  enum: [sequence, optimize]
                  default: sequence
                  description: Processing mode
                sizing_profile_id:
                  type: integer
                  description: Optional default sizing profile ID
                machine_id:
                  type: integer
                  description: Optional target machine ID
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      tags:
        - jobs
      summary: List jobs with pagination
      operationId: listJobs
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, processing, needs_input, completed, failed]
      responses:
        '200':
          description: Paginated list of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
  /v1/jobs/{job_id}:
    get:
      tags:
        - jobs
      summary: Get job details
      description: Get detailed information about a job including items
      operationId: getJobDetail
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - jobs
      summary: Delete/cancel job
      description: Delete a job. Only queued or failed jobs can be deleted.
      operationId: deleteJob
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/JobId'
      responses:
        '204':
          description: Job deleted successfully
        '400':
          description: Job cannot be deleted (e.g., currently processing)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/jobs/{job_id}/pending-items:
    get:
      tags:
        - jobs
      summary: Get pending items
      description: |
        Get items that need manual resolution (missing or ambiguous SKUs).
        Returns items with status 'missing' or 'ambiguous'.
        For ambiguous items, includes candidate assets with similarity scores.
      operationId: getPendingItems
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: List of pending items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingItemsResponse'
        '400':
          description: Job is not in 'needs_input' status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/jobs/{job_id}/resolve:
    post:
      tags:
        - jobs
      summary: Resolve pending items manually
      description: |
        Manually resolve items by assigning asset IDs.
        After resolution, if all items are resolved, job processing continues automatically.
      operationId: resolveItems
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobResolveRequest'
      responses:
        '200':
          description: Items resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResolveResponse'
        '400':
          description: Invalid resolution (e.g., asset not found, item already resolved)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/jobs/{job_id}/skip:
    post:
      tags:
        - jobs
      summary: Skip pending items
      description: |
        Skip items that don't have matching images. These items will be excluded from the generated base.
        After skipping, if all items are resolved or skipped, job processing continues automatically.
        
        **Use cases:**
        - Item SKU not found in asset catalog
        - User chooses to generate base without this item
        - Item will be marked as "skipped" and excluded from packing
      operationId: skipItems
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSkipRequest'
      responses:
        '200':
          description: Items skipped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobSkipResponse'
        '400':
          description: Invalid request (e.g., job not in needs_input state, item not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/jobs/{job_id}/outputs:
    get:
      tags:
        - jobs
      summary: Get job output PDFs
      description: Get list of generated PDF files for completed jobs
      operationId: getJobOutputs
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: List of output PDFs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOutputsResponse'
        '400':
          description: Job is not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant ID for multi-tenancy
      schema:
        type: integer
...
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
...
    JobCreateResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [queued, processing, needs_input, completed, failed]
        mode:
          type: string
          enum: [sequence, optimize]
        picklist_uri:
          type: string
        machine_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
    JobDetailResponse:
      type: object
      properties:
        id:
          type: integer
        tenant_id:
          type: integer
        machine_id:
          type: integer
          nullable: true
        sizing_profile_id:
          type: integer
          nullable: true
        status:
          type: string
        mode:
          type: string
        picklist_uri:
          type: string
        manifest_json:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        items_count:
          type: integer
        items_resolved:
          type: integer
        items_pending:
          type: integer
        items_skipped:
          type: integer
    JobListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              status:
                type: string
              mode:
                type: string
              picklist_uri:
                type: string
              items_count:
                type: integer
              created_at:
                type: string
                format: date-time
              updated_at:
                type: string
                format: date-time
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer
    PendingItemsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              sku:
                type: string
              quantity:
                type: integer
              size_label:
                type: string
                nullable: true
              status:
                type: string
                enum: [missing, ambiguous]
              candidates:
                type: array
                items:
                  type: object
                  properties:
                    asset_id:
                      type: integer
                    sku:
                      type: string
                    file_uri:
                      type: string
                    score:
                      type: number
                      format: float
    JobResolveRequest:
      type: object
      required:
        - resolutions
      properties:
        resolutions:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - item_id
              - asset_id
            properties:
              item_id:
                type: integer
              asset_id:
                type: integer
    JobResolveResponse:
      type: object
      properties:
        status:
          type: string
        resolved_count:
          type: integer
        job_status:
          type: string
        message:
          type: string
    JobSkipRequest:
      type: object
      required:
        - item_ids
      properties:
        item_ids:
          type: array
          minItems: 1
          items:
            type: integer
          description: List of item IDs to skip
    JobSkipResponse:
      type: object
      properties:
        status:
          type: string
        skipped_count:
          type: integer
        job_status:
          type: string
        message:
          type: string
    JobOutputsResponse:
      type: object
      properties:
        pdfs:
          type: array
          items:
            type: string
            description: URI of PDF file in storage
        previews:
          type: array
          items:
            type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...